fix: cant deleted accosiated data

+++ app/Http/Controllers/apps/AreaController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\AreaStoreRequest;
 use App\Http\Requests\AreaUpdateRequest;
 use App\Models\Area;
 use App\Models\Region;
+use App\Models\TransporterRate;
 use Illuminate\Http\Request;
 
 class AreaController extends Controller
@@ -49,27 +50,32 @@ class AreaController extends Controller
 
   public function update(AreaUpdateRequest $request, $uuid)
   {
-      $validated = $request->validated();
-      $findArea = Area::where('ShortChar01', $validated['code'])->first();
-      $area = Area::findOrFail($uuid);
+    $validated = $request->validated();
+    $findArea = Area::where('ShortChar01', $validated['code'])->first();
+    $area = Area::findOrFail($uuid);
 
-      if($findArea and $area->uuid != $findArea->uuid) {
-          return redirect()->route(
-              'master-data.area.edit',
-              ['uuid' => $uuid]
-          )->with('error', 'Code already exist!');
-      }
-
-      $area->update($validated);
+    if ($findArea and $area->uuid != $findArea->uuid) {
       return redirect()->route(
-          'master-data.area.edit',
-          ['uuid' => $uuid]
-      )->with('success', 'Area updated successfully');
+        'master-data.area.edit',
+        ['uuid' => $uuid]
+      )->with('error', 'Code already exist!');
+    }
+
+    $area->update($validated);
+    return redirect()->route(
+      'master-data.area.edit',
+      ['uuid' => $uuid]
+    )->with('success', 'Area updated successfully');
   }
 
   public function delete($uuid)
   {
     $area = Area::findOrFail($uuid);
+
+    $transporterRate = TransporterRate::where('Key2', $uuid)->exists();
+    if ($transporterRate) {
+      return redirect()->route('master-data.area.list')->with('error', 'Failed to delete Area. Reason: The Area is already associated with a Transporter Rate`s.');
+    }
     $area->delete();
 
     return redirect()->route('master-data.area.list')->with('success', 'Area deleted successfully.');
